新版轉折點判斷
1.可以連續平點之後的轉折。
2.交易紀錄增加index

+++ QuickTradeTest/OriginalRecordConverter.cs
@@ -55,7 +55,7 @@ namespace QuickTradeTest
 
                 tradeCurrentVolumn = int.Parse(tmpData[5].Trim());
 
-                record.TradePrice = Convert.ToDouble(currentPrice);
+                record.TradePrice = Convert.ToInt32(currentPrice);
 
                 record.TradeVolumn = tradeCurrentVolumn / 2;
